Added stat comparisons for height, turnovers, personalfouls
