+++ .eslintignore
@@ -1,7 +1,5 @@
 node_modules
-sql/seeds
+scripts/seeds
 models/index.js
 config/config.js
-gamePage/page.js
-loginPage/login.js
 .env
