+++ config/config.js
@@ -1,4 +1,3 @@
-/* eslint-disable */
 require("dotenv").config();
 
 const {
