--- routes/html-routes.js
@@ -1,7 +1,0 @@
-const path = require("path");
-
-module.exports = function (app) {
-  app.get("/", async (req, res) => {
-    res.sendFile(path.join(__dirname, "../public/index.html"));
-  });
-};
