+++ scripts/seedBomb.js
