+++ server.js
@@ -3,9 +3,10 @@ const session = require("express-session");
 const compression = require("compression");
 const mongoose = require("mongoose");
 const morgan = require("morgan");
+const path = require("path");
 const passport = require("./config/passport");
 const db = require("./models");
-// const SeedBomb = require("./sql/seedBomb");
+// const SeedBomb = require("./scripts/seedBomb");
 
 const PORT = process.env.PORT || 3001;
 const app = express();
@@ -16,21 +17,25 @@ app.use(morgan("dev"));
 app.use(express.urlencoded({ extended: true }));
 app.use(express.json());
 
+// Define API routes here
+app.use("/auth", require("./routes/auth"));
+app.use("/api/game", require("./routes/game"));
+
 // Serve up static assets (usually on heroku)
 if (process.env.NODE_ENV === "production") {
   app.use(express.static("client/build"));
 }
 
 // We need to use sessions to keep track of our user's login status
-app.use(
-  session({ secret: "keyboard cat", resave: true, saveUninitialized: true }),
-);
+app.use(session({ secret: "where is hunter", resave: true, saveUninitialized: true }));
 app.use(passport.initialize());
 app.use(passport.session());
 
-// Define API routes here
-require("./routes/api-routes.js")(app);
-require("./routes/html-routes.js")(app);
+// Send every other request to the React app
+// Define any API routes before this runs
+app.get("*", (req, res) => {
+  res.sendFile(path.join(__dirname, "./client/build/index.html"));
+});
 
 // Setup game state in nosql
 mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/wargame", {
@@ -40,6 +45,8 @@ mongoose.connect(process.env.MONGODB_URI || "mongodb://localhost/wargame", {
   useCreateIndex: true,
 }).then(() => {
   console.log("Mongo Database Connected");
+}).catch((error) => {
+  if (error) throw error;
 });
 
 // Syncing our database and logging a message to the user upon success
